+++ src/Query.hs
@@ -36,16 +36,15 @@ pkgAttrsOptions =
     "supplements"
   ]
 
--- FIXME --no-redirect?
 -- FIXME error if no testing repo
 repoqueryCmd :: Bool -> Bool -> Verbosity -> Bool -> Bool -> Bool -> Release
              -> RepoSource -> Arch -> [Arch] -> Bool -> [String] -> IO ()
-repoqueryCmd dnf4 debug verbose multiple redirect checkdate release reposource sysarch archs testing args = do
+repoqueryCmd dnf4 debug verbose multiple dynredir checkdate release reposource sysarch archs testing args = do
   forM_ (if null archs then [sysarch] else archs) $ \arch -> do
     repoConfigs <-
       if release == System
       then return []
-      else showRelease debug redirect True checkdate reposource release sysarch (Just arch) testing
+      else showRelease debug dynredir True checkdate reposource release sysarch (Just arch) testing
     let qfAllowed = not $ any (`elem` ["-i","--info","-l","--list","-s","--source","--nvr","--nevra","--envra","--qf","--queryformat", "--changelog"] ++ pkgAttrsOptions) args
     -- dnf5 writes repo update output to stdout
     -- https://github.com/rpm-software-management/dnf5/issues/1361