+++ ChangeLog.md
@@ -3,9 +3,9 @@
 ## 0.6 (2024-07-16)
 - defaults to no datestamp output: --time replaces --no-check
 - further speed improvements by reducing http checks
-- add --redirect now to enable redirect re-use
+- add --dynamic redirect to not re-use first mirror redirect
 - update c8s url to Centos Vault
-- replace spanJust with simpler partitionEithers
+- drop dep on utility-ht just for spanJust
 
 ## 0.5 (2024-07-02)
 - allow multiple release args